% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estiParamSingle.R
\name{estiParamSingle}
\alias{estiParamSingle}
\title{Parameter estimation for scDNAm data under the 1-group scenario.}
\usage{
estiParamSingle(scDNAm_mat, ptime)
}
\arguments{
\item{scDNAm_mat}{A numeric matrix containing the single-cell DNA methylation level. Methylation levels
can be calculated as the ratio of methylation coverage over the total coverage. Genomic feature (gene) names
are labeled in rownames, and cells are labeled in colnames.}

\item{ptime}{A numeric vector containing the estimated pseudotime of all cells in "scDNAm_mat".}
}
\value{
A numeric list of estimated parameters for all genomic features that will be used as the input
of the function \code{\link{dmSingle}}. Parameters include:
\itemize{
\item \eqn{\beta_0}: The estimated intercept of the polynomial of degree 4.
\item \eqn{\beta_1}: The estimated coefficient for the linear term of the polynomial.
\item \eqn{\beta_2}: The estimated coefficient for the quadratic term of the polynomial.
\item \eqn{\beta_3}: The estimated coefficient for the cubic term of the polynomial.
\item \eqn{\beta_4}: The estimated coefficient for the quartic term of the polynomial.
\item \eqn{\sigma^2_1}: The estimated variance for the first stage along the pseudotime.
\item \eqn{\sigma^2_2}: The estimated variance for the second stage along the pseudotime.
\item \eqn{\sigma^2_3}: The estimated variance for the third stage along the pseudotime.
\item \eqn{\sigma^2_4}: The estimated variance for the fourth stage along the pseudotime.
}
}
\description{
This function performs the Gibbs sampling procedure based on the hierarchical Bayesian modeling,
to produce the parameters required in the differential methylation analysis.
}
