% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fixdiag.R
\name{fixdiag}
\alias{fixdiag}
\title{Fix diagonal elements}
\usage{
fixdiag(Ncol, Nrow)
}
\description{
\code{fixdiag} generates an index-matrix for a symmetrical matrix with \code{NA} along the diagonal
}

