% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiST.data.R
\docType{data}
\name{MiST.data}
\alias{MiST.data}
\title{MiST.data}
\format{A [\link{list}] object.}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE74821}{GSE74821}
}
\usage{
MiST.data
}
\description{
MiST.data
}
\details{
\code{MiST.data} contains:
\itemize{
\item \code{G} a numeric genotype matrix of 3400 individuals and 93 SNPs. Each row represents a individual, and each column represents a SNP marker.
\item \code{X} a numeric matrix of covariates with the first column represents intercept, the second column represents a continuous covariate, and the third column represents a binary covariate.
\item \code{Z} a numeric vector of 1s for the second level covariate.
\item \code{Z.func} a numeric matrix for the second level covariate. The first column contains all 1s, representing the intercept, and the second column is 0 or 1, representing whether a SNP marker is functional or non-functional.
\item \code{maf} a numeric vector for the Minor Allele Frequecy of 93 SNPs.
\item \code{y.con} a numeric vector of continuous outcomes.
\item \code{y.bin} a numeric vector of binary outcomes.
}
}
\keyword{datasets}
