% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mist_data.R
\docType{data}
\name{mist_data}
\alias{mist_data}
\title{mist_data}
\format{
A [\link{list}] object.
}
\usage{
mist_data
}
\description{
mist_data
}
\details{
\code{mist_data} contains:
\itemize{
\item \code{phenotypes} a data.frame with outcomes (continuous and binary) and covariates.
\item \code{G} a numeric matrix of genotypes
\item \code{Z} a numeric matrix with information on variants, \emph{i.e.}, group, maf and effect.
}
}
\keyword{datasets}
