% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{plot_interactions}
\alias{plot_interactions}
\title{Plot interaction matrix}
\usage{
plot_interactions(Report, TmbData, category_names = 1:TmbData$n_p,
  plotdir = paste0(getwd(), "/"), ...)
}
\description{
\code{plot_interactions} plots the estimated interaction matrix
}

