% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene.R
\name{plotGene}
\alias{plotGene}
\title{Plot Methylation Levels vs. Pseudotime for a Specific Gene}
\usage{
plotGene(Dat_sce, Dat_name, ptime_name, beta_sigma_list, gene_name)
}
\arguments{
\item{Dat_sce}{A \code{SingleCellExperiment} object containing the single-cell DNA methylation level.
Methylation levels should be stored as assays, with genomic feature (gene) names in rownames
and cells in colnames.}

\item{Dat_name}{A character string specifying the name of the assay to extract the methylation data.}

\item{ptime_name}{A character string specifying the name of the colData to extract the pseudotime vector.}

\item{beta_sigma_list}{A named numeric list of estimated coefficients (output of \code{estiParamSingle}).}

\item{gene_name}{A character string specifying the gene name to plot.}
}
\value{
A ggplot2 scatter plot with an overlayed fitted curve.
}
\description{
Generates a scatter plot of methylation levels vs. pseudotime for a specific gene,
overlaid with the fitted curve based on the estimated coefficients.
}
\examples{
\dontrun{
plotGene(Dat_sce = Dat_sce,
Dat_name = "Methy_level_group1",
ptime_name = "pseudotime",
beta_sigma_list,
gene_name = "ENSMUSG00000000037")
}
}
