% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ptime_GSE121708}
\alias{ptime_GSE121708}
\title{Pseudotime Data for GSE121708 Dataset}
\format{
A numeric vector representing pseudotime values for cells.
}
\source{
GSE121708 dataset from the GEO database.
}
\usage{
ptime_GSE121708
}
\description{
This dataset contains pseudotime information for single-cell DNA methylation data from the GSE121708 study.
}
\keyword{datasets}
