% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{scDNAm_GSE121708}
\alias{scDNAm_GSE121708}
\title{Single-cell DNA Methylation Data for GSE121708}
\format{
A matrix with rows representing genomic features and columns representing cells. The values are methylation levels (e.g., ratios).
}
\source{
GSE121708 dataset from the GEO database.
}
\usage{
scDNAm_GSE121708
}
\description{
This dataset contains single-cell DNA methylation data from the GSE121708 study. The data includes methylation levels for genomic features across cells.
}
\keyword{datasets}
