% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{scDNAm_GSE121708_g2}
\alias{scDNAm_GSE121708_g2}
\title{Group 2 Single-cell DNA Methylation Data for GSE121708}
\format{
A matrix with rows representing genomic features and columns representing cells. The values are methylation levels (e.g., ratios) for the second group.
}
\source{
GSE121708 dataset from the GEO database.
}
\usage{
scDNAm_GSE121708_g2
}
\description{
This dataset contains single-cell DNA methylation data for the second group from the GSE121708 study.
}
\keyword{datasets}
