% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sim_Fn.R
\name{Sim_Fn}
\alias{Sim_Fn}
\title{Simulate data}
\usage{
Sim_Fn(n_species = 4, n_years = 20, n_years_burnin = 0, n_stations = 25,
  n_samp_per_station = 1, n_knots = n_stations,
  start_from_equilibrium = TRUE, B_pp = NULL, ObsModel = "Poisson",
  Cov_pp = NULL, B_params = c(0, 0.2), phi_p = NULL, sdlog = 0.1,
  SpatialScale = 0.1, SD_A = 0.5, SD_E = 0.2, corr_E = 0.5, rho = 0.8,
  logMeanDens = 1, RandomSeed = NA, Loc = NULL)
}
\description{
\code{Sim_Fn} simulates data from the data-generating process used in the SpatialVAM model
}

