% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmSingle.R
\name{dmSingle}
\alias{dmSingle}
\title{Differential methylation evaluation over time for scDNA-seq data under the 1-group scenario.}
\usage{
dmSingle(beta_sigma_list)
}
\arguments{
\item{beta_sigma_list}{A numeric list generated by the function \code{estiParamSingle}. Each element of the list corresponds to a genomic feature (e.g., a gene) and contains a numeric vector of length 9 with estimated values: \eqn{\beta_0}, \eqn{\beta_1}, \eqn{\beta_2}, \eqn{\beta_3}, \eqn{\beta_4}, \eqn{\sigma^2_1}, \eqn{\sigma^2_2}, \eqn{\sigma^2_3}, \eqn{\sigma^2_4}.}
}
\value{
A named numeric vector where each value corresponds to a genomic feature
(e.g., a gene). The values represent the minimum area between the fitted curve
of scDNA methylation levels along pseudotime and a constant horizontal line for
each feature. This metric measures the magnitude of methylation level changes
along pseudotime. The values are sorted in descending order, with larger values
indicating more drastic changes.
}
\description{
This function performs DM analysis to identify genomic features showing drastic
changes along pseudotime, by modeling the methylation changes along pseudotime and estimating
the developmental-stage-specific biological variations.
}
\examples{
library(mist)
data <- readRDS(system.file("extdata", "small_sampleData_sce.rds", package = "mist"))
beta_sigma_list <- estiParamSingle(
    Dat_sce = data,
    Dat_name = "Methy_level_group1",
    ptime_name = "pseudotime"
)
dm_results <- dmSingle(beta_sigma_list)
}
