% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmTwo.R
\name{dmTwoGroups}
\alias{dmTwoGroups}
\title{Differential methylation evaluation over time for scDNA-seq data under the 2-group scenario.}
\usage{
dmTwoGroups(beta_sigma_list_group)
}
\arguments{
\item{beta_sigma_list_group}{A list containing two elements:
\itemize{
\item \code{Group1}: A numeric list of estimated parameters for all genomic features in group 1 (generated by \code{estiParamTwoGroups}).
\item \code{Group2}: A numeric list of estimated parameters for all genomic features in group 2.
}}
}
\value{
A named numeric vector where each value corresponds to a genomic feature
(e.g., a gene). The values represent the integral of the differences between the
fitted curves of scDNA methylation levels for the two groups. The values are sorted
in descending order, with larger values indicating more drastic differences between the groups.
}
\description{
This function performs DM analysis to identify genomic features showing drastic
changes between two groups along pseudotime, by modeling the methylation changes
and comparing the fitted curves for each group.
}
\examples{
library(mist)
Dat_path <- system.file("extdata", "small_sampleData_sce.rds", package = "mist")
beta_sigma_list_group <- estiParamTwo(
    Dat_sce = Dat_path,
    Dat_name_g1 = "Methy_level_group1",
    Dat_name_g2 = "Methy_level_group2",
    ptime_name_g1 = "pseudotime",
    ptime_name_g2 = "pseudotime_g2"
)
dm_results_two <- dmTwoGroups(beta_sigma_list_group)
}
