% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmTwoGroups.R
\name{dmTwoGroups}
\alias{dmTwoGroups}
\title{Differential methylation evaluation over time for scDNA-seq data under the 2-group scenario.}
\usage{
dmTwoGroups(Dat_sce)
}
\arguments{
\item{Dat_sce}{The updated sce object with 2 numeric matrices in the rowData:
\itemize{
\item \code{mist_pars_group1}: A numeric matrix of estimated parameters for all genomic features in group 1 (generated by \code{estiParamTwoGroups}).
\item \code{mist_pars_group2}: A numeric matrix of estimated parameters for all genomic features in group 2.
}}
}
\value{
The updated sce object with a named numeric vector where each value corresponds to a genomic feature
(e.g., a gene). The values represent the integral of the differences between the
fitted curves of scDNA methylation levels for the two groups. The values are sorted
in descending order, with larger values indicating more drastic differences between the groups.
}
\description{
This function performs DM analysis to identify genomic features showing drastic
changes between two groups along pseudotime, by modeling the methylation changes
and comparing the fitted curves for each group.
}
\examples{
library(mist)
data <- readRDS(system.file("extdata", "small_sampleData_sce.rds", package = "mist"))
Dat_sce_new <- estiParamTwo(
    Dat_sce = data,
    Dat_name_g1 = "Methy_level_group1",
    Dat_name_g2 = "Methy_level_group2",
    ptime_name_g1 = "pseudotime",
    ptime_name_g2 = "pseudotime_g2"
)
dm_sce_two <- dmTwoGroups(Dat_sce_new)
}
