% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estiParamSingle.R
\name{estiParamSingle}
\alias{estiParamSingle}
\title{Parameter estimation for scDNAm data under the 1-group scenario.}
\usage{
estiParamSingle(Dat_sce, Dat_name, ptime_name)
}
\arguments{
\item{Dat_sce}{A character string specifying the file path to the SingleCellExperiment object containing
the single-cell DNA methylation level. Methylation levels can be calculated as the ratio of methylation
coverage over the total coverage. Genomic feature (gene) names are labeled in rownames, and cells are
labeled in colnames.}

\item{Dat_name}{A character string specifying the name of the assay to extract the methylation level data from
the SCE obejct. Data matrix should be stored in the 'Assay'.}

\item{ptime_name}{A character string specifying the name of the colData to extract the pseudotime vector from
the SCE obejct. Pseudotime should be stored in the 'colData'.}
}
\value{
A numeric list of estimated parameters for all genomic features that will be used as the input
of the function \code{\link{dmSingle}}. Parameters include:
\itemize{
\item \eqn{\beta_0}: The estimated intercept of the polynomial of degree 4.
\item \eqn{\beta_1}: The estimated coefficient for the linear term of the polynomial.
\item \eqn{\beta_2}: The estimated coefficient for the quadratic term of the polynomial.
\item \eqn{\beta_3}: The estimated coefficient for the cubic term of the polynomial.
\item \eqn{\beta_4}: The estimated coefficient for the quartic term of the polynomial.
\item \eqn{\sigma^2_1}: The estimated variance for the first stage along the pseudotime.
\item \eqn{\sigma^2_2}: The estimated variance for the second stage along the pseudotime.
\item \eqn{\sigma^2_3}: The estimated variance for the third stage along the pseudotime.
\item \eqn{\sigma^2_4}: The estimated variance for the fourth stage along the pseudotime.
}
}
\description{
This function performs the Gibbs sampling procedure based on the hierarchical Bayesian modeling,
to produce the parameters required in the differential methylation analysis.
}
\examples{
library(mist)
Dat_path <- system.file("extdata", "small_sampleData_sce.rds", package = "mist")
beta_sigma_list <- estiParamSingle(
    Dat_sce = Dat_path,
    Dat_name = "Methy_level_group1",
    ptime_name = "pseudotime"
)
}
