% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estiParamTwo.R
\name{estiParamTwo}
\alias{estiParamTwo}
\title{Parameter estimation for scDNAm data under the 2-group scenario.}
\usage{
estiParamTwo(scDNAm_mat_group1, scDNAm_mat_group2, ptime_group1, ptime_group2)
}
\arguments{
\item{scDNAm_mat_group1}{A numeric matrix containing the single-cell DNA methylation levels for Group 1.
Genomic feature (gene) names are labeled in rownames, and cells are labeled in colnames.}

\item{scDNAm_mat_group2}{A numeric matrix containing the single-cell DNA methylation levels for Group 2.
Genomic feature (gene) names are labeled in rownames, and cells are labeled in colnames.}

\item{ptime_group1}{A numeric vector containing the estimated pseudotime of all cells in Group 1.}

\item{ptime_group2}{A numeric vector containing the estimated pseudotime of all cells in Group 2.}
}
\value{
A list containing two elements, one for each group, where each element is a numeric list of estimated
parameters for all genomic features. Each element includes:
\itemize{
\item \eqn{\beta_0}, \eqn{\beta_1}, \eqn{\beta_2}, \eqn{\beta_3}, \eqn{\beta_4}: The estimated coefficients for
the polynomial of degree 4.
\item \eqn{\sigma^2_1}, \eqn{\sigma^2_2}, \eqn{\sigma^2_3}, \eqn{\sigma^2_4}: The estimated variances for each
stage along the pseudotime.
}
}
\description{
This function performs the Gibbs sampling procedure based on hierarchical Bayesian modeling,
separately for two groups, to produce the parameters required in the differential methylation analysis.
}
