% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Summarize.R
\name{Summarize}
\alias{Summarize}
\title{Summarize model outputs}
\usage{
Summarize(Report, TmbData = NULL, SD = NULL, Map = NULL, SimList = NULL,
  species_names = NULL)
}
\description{
\code{Summarize} calculates useful outputs
}

